Moved fraction to sepratae file
