+++ Assets/Scripts/Fraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54e3fce1b9b8a69479c4854dfe2c80c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
